cmd/vla/requirements.txt
torch>=2.2,<2.6
transformers==4.40.1
viam-sdk
tokenizers==0.19.1
timm==0.9.10
huggingface_hub<0.24
peft==0.11.1
# convert_to_lerobot.py + train_smolvla.py
pandas
pyarrow
# lerobot installs separately to pick up SmolVLA extras:
#   pip install "lerobot[smolvla] @ git+https://github.com/huggingface/lerobot.git"
# OpenVLA's pin. Newer accelerate (>=0.31) regressed the bnb 4-bit path:
# dispatch_model() calls .to() unconditionally on single-device bnb models,
# which bnb refuses with "ValueError: .to is not supported for 4-bit ...".
# 0.27.2 still has the `is_bnb_quantized` branch that forces hooks instead.
accelerate==0.27.2
pillow
tqdm
bitsandbytes; sys_platform != "darwin"
